requirements.txt
google-adk
litellm
google-generativeai
mcp==1.9.1
streamlit